requirements.txt
fastapi>=0.104.0
uvicorn>=0.30.0
python-multipart>=0.0.9
SQLAlchemy>=2.0.0
Pydantic>=2.0.0
Jinja2>=3.1.0
Pillow>=10.0.0
tensorflow>=2.16.0
pytest>=7.0.0